backend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
